Resolved BTRunAction and BTPrimaryGenerator
