+++ src/BTPrimaryGeneratorAction.cpp
@@ -71,8 +71,9 @@ void BTPrimaryGeneratorAction::GeneratePrimaries(G4Event* anEvent)
         if (envLV) fEnvelopeBox = dynamic_cast<G4Box*>(envLV->GetSolid());
     }
 
-    const G4double x0 = size * envSizeXY * (G4UniformRand() 1);
-    const G4double yo = size * envSizeXY * (G4UniformRand() 1);
+    G4double size = 0.8; 
+    const G4double x0 = size * envSizeXY * (G4UniformRand()-1); // Assuming - 1
+    const G4double y0 = size * envSizeXY * (G4UniformRand()-1);
     const G4double z0 = 1 * envSizeZ;
 
     fParticleGun->SetParticlePosition(G4ThreeVector(x0, y0, z0));
